Refactor: project properies
